feat: Add support for new messages count and counter activation
This commit adds new properties to the SupportTicket entity: new_messages_count and is_counter_active. It also updates the TicketCommentService to increment the new_messages_count when a new comment is added, and reset it to 0 when an admin views the ticket. This feature will help users keep track of new messages in their support tickets.