feat: Add reActiveCounter endpoint to SupportTicketController
This commit adds a new endpoint `reActiveCounter` to the `SupportTicketController` class. The endpoint is accessible only to users with the `ADMIN` role. It allows the reactivation of the counter for a specific ticket by setting the `is_counter_active` property to `true` and resetting the `new_messages_count` to 0. This feature will help administrators keep track of new messages in support tickets.