feat: Add endpoint for listing modules
The code changes include adding a new endpoint in the `EmployeeController` to list the modules. This endpoint is implemented in the `listModules` method, which retrieves the modules from the `EmployeeService` and returns them as a `SamModuleResponse` object. This change allows the frontend to display the list of modules available to the user.